# CheckCheck/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CheckCheck REST API",
    "version": "0.1.dev47+ga91abbb.d20260618"
  },
  "paths": {
...
    "/api/checklist": {
      "get": {
        "tags": [
          "Checklist"
        ],
        "summary": "List Checklists",
        "description": "List all CheckLists of the current user with their positions and configuration. This is a rather expensive endpoint and should be only used when really needed.",
        "operationId": "list_checklists_api_checklist_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Archived"
            }
          },
          {
            "name": "label_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specify the starting point for result sets/list",
              "default": 0,
              "title": "Offset"
            },
            "description": "Specify the starting point for result sets/list"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specify the max amount of result items",
              "default": 100,
              "title": "Limit"
            },
            "description": "Specify the max amount of result items"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CheckListApiWithSubObj_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Checklist"
        ],
        "summary": "Create Checklist",
        "description": "Create a new CheckList.",
        "operationId": "create_checklist_api_checklist_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckListApiCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListApiWithSubObj"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checklist/{checklist_id}": {
      "get": {
        "tags": [
          "Checklist"
        ],
        "summary": "Get Checklist",
        "description": "Update existing CheckList",
        "operationId": "get_checklist_api_checklist__checklist_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListApiWithSubObj"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Checklist"
        ],
        "summary": "Update Checklist",
        "description": "Update existing CheckList",
        "operationId": "update_checklist_api_checklist__checklist_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckListUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListApiWithSubObj"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Checklist"
        ],
        "summary": "Delete Checklist",
        "description": "Delete existing CheckList. The respective users specific CheckListPosition will be deleted as well.",
        "operationId": "delete_checklist_api_checklist__checklist_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckListApiCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The display name of the list"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "A text that will be shown at the header"
          },
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "checked_items_seperated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Seperated",
            "default": true
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If checked_items_seperated is enabled and this is set to true the client should hide any checked items. ",
            "default": true
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CheckListPositionApiCreate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CheckListApiCreate"
      },
      "CheckListApiWithSubObj": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The display name of the list"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "A text that will be shown at the header"
          },
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "checked_items_seperated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Seperated",
            "default": true
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If checked_items_seperated is enabled and this is set to true the client should hide any checked items. ",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChecklistColorScheme"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "$ref": "#/components/schemas/CheckListPositionPublicWithoutChecklistID"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/LabelReadAPI"
            },
            "type": "array",
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "id",
          "color",
          "position",
          "labels"
        ],
        "title": "CheckListApiWithSubObj"
      },
...
      "CheckListPositionApiCreate": {
        "properties": {
          "index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index",
            "description": "The position of a checklist relative to other checklist 'index' value. The lowest value should be the first checklist in a client representation."
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned",
            "description": "Pinned items should get a more prominent placement in the client.",
            "default": false
          },
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived",
            "description": "Archived items are filtered out by default. They are considered obsolete.",
            "default": false
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If true the client should hide checked items in the list view.",
            "default": false
          }
        },
        "type": "object",
        "title": "CheckListPositionApiCreate"
      },
      "CheckListPositionPublicWithoutChecklistID": {
        "properties": {
          "index": {
            "type": "number",
            "title": "Index",
            "description": "The position of a checklist relative to other checklist 'index' value. The lowest value should be the first checklist in a client representation."
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned",
            "description": "Pinned items should get a more prominent placement in the client.",
            "default": false
          },
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived",
            "description": "Archived items are filtered out by default. They are considered obsolete.",
            "default": false
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If true the client should hide checked items in the list view.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "index"
        ],
        "title": "CheckListPositionPublicWithoutChecklistID"
      },
...
      "CheckListUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The display name of the list"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "A text that will be shown at the header"
          },
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "checked_items_seperated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Seperated",
            "default": true
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If checked_items_seperated is enabled and this is set to true the client should hide any checked items. ",
            "default": true
          }
        },
        "type": "object",
        "title": "CheckListUpdate"
      },
...
      "ChecklistColorScheme": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Name of the color."
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "description": "Reserverd for later use. Color order not implemented yet"
          },
          "backgroundcolor_dark_hex": {
            "type": "string",
            "title": "Backgroundcolor Dark Hex"
          },
          "backgroundcolor_light_hex": {
            "type": "string",
            "title": "Backgroundcolor Light Hex"
          },
          "textcolor_dark_hex": {
            "type": "string",
            "title": "Textcolor Dark Hex"
          },
          "textcolor_light_hex": {
            "type": "string",
            "title": "Textcolor Light Hex"
          },
          "accentcolor_dark_hex": {
            "type": "string",
            "title": "Accentcolor Dark Hex"
          },
          "accentcolor_light_hex": {
            "type": "string",
            "title": "Accentcolor Light Hex"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "backgroundcolor_dark_hex",
          "backgroundcolor_light_hex",
          "textcolor_dark_hex",
          "textcolor_light_hex",
          "accentcolor_dark_hex",
          "accentcolor_light_hex"
        ],
        "title": "ChecklistColorScheme"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelReadAPI": {
        "properties": {
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Name of the label."
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "description": "Label order per user"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChecklistColorScheme"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "color"
        ],
        "title": "LabelReadAPI"
      },
...
      "PaginatedResponse_CheckListApiWithSubObj_": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of items in the database.",
            "examples": [
              300
            ]
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset",
            "description": "Starting position index of the returned items in the dataset.",
            "default": 0,
            "examples": [
              299
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of items returned in the response",
            "examples": [
              1
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CheckListApiWithSubObj"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items returned in the response following given criteria"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "count",
          "items"
        ],
        "title": "PaginatedResponse[CheckListApiWithSubObj]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}